Debug ver 0.4 (stuff happened)
*Added facebook Tests to Rep
(Input files had to be corrected, so we want to keep them in the Rep).

*xor_mask wasn't calculated correctly AT ALL.
changed it to unsigned (cause DUH)
moved it to be global, and set it at INIT. realised it's exactly like HR_mask.
Almost positively can remove it.

*All tests that fail are "using_share_lsb".
I suspect the xor doesn't work  correctly., I'm clever that way.